refactor valkey integration for separation of concerns
