+++ ddtrace/contrib/internal/valkey/asyncio_patch.py
@@ -1,7 +1,7 @@
 from ddtrace import config
-from ddtrace._trace.utils_valkey import _instrument_valkey_cmd
-from ddtrace._trace.utils_valkey import _instrument_valkey_execute_async_cluster_pipeline
-from ddtrace._trace.utils_valkey import _instrument_valkey_execute_pipeline
+from ddtrace.contrib.internal.valkey.patch import _instrument_valkey_cmd
+from ddtrace.contrib.internal.valkey.patch import _instrument_valkey_execute_async_cluster_pipeline
+from ddtrace.contrib.internal.valkey.patch import _instrument_valkey_execute_pipeline
 from ddtrace.contrib.internal.valkey_utils import _run_valkey_command_async
 from ddtrace.internal.utils.formats import stringify_cache_args
